requirements.txt
crewai
crewai-tools
python-dotenv
yfinance
# The following packages prevent the LiteLLM red error logs
fastapi
uvicorn
apscheduler